move fake_cloud_server to demo
