+++ homematicip_demo/json_data/unknown_types.json
