+++ tests/conftest.py
@@ -1,12 +1,4 @@
-from pathlib import Path
-
 import pytest
-import json
-
-from functools import partialmethod, partial
-import warnings
-import requests
-import contextlib
 
 from homematicip.aio.auth import AsyncAuth
 from homematicip.aio.home import AsyncHome
@@ -14,37 +6,11 @@ from homematicip.home import Home
 from homematicip.connection import Connection
 
 from pytest_localserver.http import WSGIServer
-from fake_cloud_server import FakeCloudServer
-
+from homematicip_demo.fake_cloud_server import FakeCloudServer
+from homematicip_demo.helper import *
 
 from datetime import datetime, timedelta, timezone
 import time
-
-
-def get_full_path(name):
-    """Returns full path of incoming relative path.
-    Relative path is relative to the script location.
-    """
-    pth = Path(__file__).parent.joinpath(name)
-    return pth
-
-
-def fake_home_download_configuration():
-    _full = get_full_path("json_data/home.json")
-    with open(_full, encoding="UTF-8") as f:
-        return json.load(f)
-
-
-@contextlib.contextmanager
-def no_ssl_verification():
-    old_request = requests.Session.request
-    requests.Session.request = partialmethod(old_request, verify=False)
-
-    warnings.filterwarnings("ignore", "Unverified HTTPS request")
-    yield
-    warnings.resetwarnings()
-
-    requests.Session.request = old_request
 
 
 @pytest.fixture
